[Gold IV] Title: 카드 정렬하기, Time: 364 ms, Memory: 26048 KB -BaekjoonHub

+++ 백준/Gold/1715. 카드 정렬하기/카드 정렬하기.java
@@ -4,17 +4,17 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        PriorityQueue<Long> pq = new PriorityQueue<>();
+        PriorityQueue<Integer> pq = new PriorityQueue<>();
 
         int N = Integer.parseInt(br.readLine());
 
         for(int i = 0 ; i < N ; i ++)
-            pq.offer(Long.parseLong(br.readLine()));
+            pq.offer(Integer.parseInt(br.readLine()));
 
         long res = 0;
         while(pq.size() > 1){
-            long a = pq.poll();
-            long b = pq.poll();
+            int a = pq.poll();
+            int b = pq.poll();
 
             res += a+b;
             pq.offer(a+b);
